\name{RcppL1TF-package}
\alias{RcppL1TF-package}
\alias{RcppL1TF}
\docType{package}
\title{\packageTitle{RcppL1TF}}
\description{\packageDescription{RcppL1TF}}
\section{Package Content}{\packageIndices{RcppL1TF}}
\author{\packageAuthor{RcppL1TF}}
\section{Maintainer}{\packageMaintainer{RcppL1TF}}
\keyword{package}
